feat: Add dynamic display mode detection with precise refresh rates
- Added /api/display-modes endpoint to query actual display capabilities
- Calculates precise refresh rates from DRM timings (60.00 vs 59.94 Hz)
- Changed refresh_hz from int to float throughout codebase for precision
- Web UI now populates refresh rate dropdown from hardware capabilities
- Distinguishes between 60.00 Hz (true 60) and 59.94 Hz (NTSC)
- Important for LTC timecode drop-frame vs non-drop-frame accuracy
- Updated config parsing with json_extract_float() function
- Added fabsf() comparison for float refresh rate change detection
- Web UI properly matches config values to available display modes

Display modes detected:
- 60.00 Hz (standard)
- 59.94 Hz (NTSC for drop-frame LTC)
- 50.00 Hz (PAL)

+++ include/config-management.h
@@ -7,7 +7,7 @@ typedef struct {
     char timezone[256];           // e.g., "Europe/London"
     char ntp_server[256];         // Currently selected NTP server
     char custom_ntp_servers[2048]; // JSON array of custom servers
-    int refresh_hz;               // 50 or 60
+    float refresh_hz;             // Display refresh rate (e.g., 50.0, 59.94, 60.0)
     int timecode_x;               // X position
     int timecode_y;               // Y position
     int color_r;                  // Red component
